dashboard: move Changelog to own file
